Move subgraphviews into graphviews
