+++ networkx/classes/__init__.py
@@ -3,7 +3,6 @@ from .digraph import DiGraph
 from .multigraph import MultiGraph
 from .multidigraph import MultiDiGraph
 from .views import *
-from .subgraphviews import *
 from .graphviews import *
 from .ordered import *
